Struct pir_8_emu::isa::GeneralPurposeRegister

pub struct GeneralPurposeRegister { /* fields omitted */ }

There are eight 8-bit General Purpose registers, each has an internal address for use within the CPU, instructions like 'MOVE' and 'LOAD' can use these addresses.

All the registers will start with an initial value of 0.

There are eight 8-bit General Purpose registers, each has an internal address for use within the CPU, instructions like 'MOVE' and 'LOAD' can use these addresses. The first five registers have some special functionality, as described, the last three have no special functionality.

The last four registers can also be used with the stack.

Address	Letter	Description

000	F	Flag register (can also be used to get a zero value)

001	S	Output of the ALU - ALU operations will overwrite any value stored

010	X	Input to ALU (Only input for unary operations)

011	Y	Second input for ALU

100	A	Port number for PORT instruction

101	B

110	C

111	D

Flag register

The flag register can be be read and written to as a general purpose register, though keep in mind that ALU and Compare instructions can effect the value of the flags.

Not all of the bits have a specified role (yet), though the CLRF operation will still clear them.

A value of 1 denotes the flag as 'set', whilst a value of 0 denotes the flag is 'unset'.

Below is a description of what each bit in the flag register denotes.

Bit	Letter	Description

0	Z	Zero flag

1	C	Carry flag

2	P	Parity (even number of set bits)

3	E	Equals flag

4	G	Greater than

5

6

7

Methods

impl GeneralPurposeRegister

pub fn new(address: u8, letter: char) -> Option<GeneralPurposeRegister>

Create a new, empty register named as specified.

Examples

let x = GeneralPurposeRegister::new(0b010, 'X').unwrap();
assert_eq!(*x, 0);
assert_eq!(x.address(), 0b010);
assert_eq!(x.letter(), 'X');

assert_eq!(GeneralPurposeRegister::new(0b1000, 'Q'), None);

pub fn address(self) -> u8

The address for this register

Limited to 3 bits' width

pub fn letter(self) -> char

The letter/mnemonic for this register

E.g. "X"

pub fn relabel(&mut self, to_letter: char) -> Result<(), ()>

Change this register's letter to the specified one

Will return Err(()) if the letter isn't ASCII

impl GeneralPurposeRegister

pub fn defaults() -> GeneralPurposeRegisterBank

Get the default 8 GP registers specified in the ISA

pub fn from_letters(s: &str) -> Result<GeneralPurposeRegisterBank, i8>

Create a GP register bank from letters contained within the specified string

The string must be 8-ASCII-characters-long.

If it's shorter, Err(-1) is returned, if it's longer, Err(8) is returned, if character for register at address AAA is non-ASCII, Err(AAA) is returned,

Examples

assert_eq!(GeneralPurposeRegister::from_letters("FSXYABCD"), Ok(GeneralPurposeRegister::defaults()));