Module pir_8_emu::microcode

Structs

DisplayMicroOp
Helper struct for human-printing μOps with format!() and {}.

Enums

MicroOp
Actual μOps executable by the CPU

MicrocodeExecutionError
An error that could've occurred when performing a μOp.

Statics

NEXT_INSTRUCTION
μOps to execute to get to the next instruction.

Type Definitions

MicroOpBlock
[MicroOp; N] typedef, where N is not to be relied upon