Struct pir_8_emu::vm::Ports

pub struct Ports { /* fields omitted */ }

256B of I/O ports with R/W tracking and per-port handler logic

Methods

impl Ports

pub fn new() -> Ports

Create fresh zero-initialised unread and unwritten ports with no handlers

pub fn install_handler<H: PortHandler + 'static>(
&mut self,
handler: H,
ports: &[u8]
) -> Result<usize, (H, PortHandlerInstallError)>

Install the specified handler on the specified ports

On success, calls PortHandler::init() on the specified handler and returns a unique ID thereof

Will return an error if the specified ports are already occupied, or the handler doesn't take the amount of ports specified, or too many handlers've been registered

Examples

struct InitableHandler(Option<u8>);
impl PortHandler for InitableHandler {
    fn init(&mut self, ports: &[u8]) { self.0 = Some(ports[0]); }
}

let mut ports = Ports::new();

let handler_id = ports.install_handler(InitableHandler(None), &[0xA1])
                      .map_err(|(_, e)| e).unwrap();

assert_eq!(ports.get_handler(handler_id).and_then(|h| h.downcast_ref()),
           Some(&InitableHandler(Some(0xA1))));

pub fn get_handler(&self, idx: usize) -> Option<&(dyn PortHandler + 'static)>

Get reference to the handler with the specified ID, if exists

pub fn get_handler_mut(
&mut self,
idx: usize
) -> Option<&mut (dyn PortHandler + 'static)>

Get mutable reference to the handler with the specified ID, if exists

pub fn uninstall_handler(
&mut self,
idx: usize
) -> Option<Box<dyn PortHandler + 'static>>

Remove and unregister the handler with the specified ID and return it, if exists

Examples

struct InitableHandler(Option<u8>);
impl PortHandler for InitableHandler {
    fn init(&mut self, ports: &[u8]) { self.0 = Some(ports[0]); }
}

let mut ports = Ports::new();

let handler_id = ports.install_handler(InitableHandler(None), &[0xA1])
                      .map_err(|(_, e)| e).unwrap();
assert_eq!(ports.uninstall_handler(handler_id).and_then(|h| h.downcast().ok()),
           Some(Box::new(InitableHandler(Some(0xA1)))));

assert!(ports.get_handler(handler_id).is_none());

pub fn iter_rw(&self) -> PortsReadWrittenIterator

Get an iterator over the read and written port cells

Examples

let mut ports = Ports::new();
ports[0x4B] = ports[0xA1];
println!("{}", ports[0x4B]);
ports[0xEB] = 0x12;

// (address, value, was_read, was_written)
assert_eq!(ports.iter_rw().collect::<Vec<_>>(),
           &[(0x4B, 0x00, true, true),
             (0xA1, 0x00, true, false),
             (0xEB, 0x12, false, true)]);

pub fn reset_rw(&mut self)

Mark all ports as unread and unwritten

Examples

let mut ports = Ports::new();
ports[0x4B] = ports[0xA1];
println!("{}", ports[0x4B]);
ports[0xEB] = 0x12;

ports.reset_rw();
assert_eq!(ports.iter_rw().collect::<Vec<_>>(), &[]);