Enum pir_8_emu::isa::instruction::Instruction

pub enum Instruction {
    Reserved(u8),
    Jump(InstructionJumpCondition),
    LoadImmediate {
        aaa: u8,
    },
    LoadIndirect {
        aaa: u8,
    },
    Save {
        aaa: u8,
    },
    Alu(AluOperation),
    Move {
        aaa: u8,
        bbb: u8,
    },
    Port {
        d: InstructionPortDirection,
        aaa: u8,
    },
    Comp {
        aaa: u8,
    },
    Stck {
        d: InstructionStckDirection,
        r: InstructionStckRegisterPair,
    },
    Clrf,
    Halt,
}

Instructions will increase the PC by one, unless otherwise stated.

The PC is incremented as the instruction is loaded from RAM. An instruction is a single byte, and can include some following immediate values purely for data.

The 'Bit Mask' shows a pattern which denotes an instruction or group of instructions, the letters denoting where any value can be used and still be considered part of the same instruction. The 'name' is for either a group or single instruction. 'Count' is how many of the 256 possible instructions are used by that bit pattern; HALT for example is exactly one instruction, whilst MOVE is effectively 64 possible combinations [this was added to help me keep track of how many operations I've defined, it should add up to 256].

Bit Mask	Name	Count	Description

0000 XXXX		16	Reserved

0001 0XXX	JUMP	8	Jump, see section below

0001 1AAA	LOAD	8	Load the the next byte into register AAA

(PC will be incremented a second time)

0010 0AAA	LOAD	8	Load value in address indicated by the next two bytes into register AAA

(PC will be incremented two more times)

0010 1AAA	SAVE	8	Store value in register AAA in address indicated by the next two bytes

(PC will be incremented two more times)

0011 XXXX	ALU	16	ALU based operations, see section below

01AA ABBB	MOVE	64	Move a value from register AAA to register BBB

10XX XXXX		64	Reserved

110X XXXX		32	Reserved

1110 XXXX	PORT	16	Perform I/O, see section below

1111 0AAA	COMP	8	Compare register S with register AAA, see section below

1111 10XX	STCK	4	Stack manipulation, see section below

1111 110X		2	Reserved

1111 1110	CLRF	1	Clear the 'F' register, by setting it to 0000 0000

1111 1111	HALT	1	Stop the CPU from doing any more execution

PORT - I/O

The PORT instruction in the form 1110 DAAA will perform I/O on the port specified in register A.

The D bit specifies the direction - 1 for reading in from the port (PORT IN) and 0 for writing out to the port (PORT OUT). The AAA bits specify the register to write to (D=1) or read from (D=0).

COMP - Compare

The compare instruction will compare the S register with a selected register. It will set the Zero and Parity flag based on the value of the S register; the Zero flag if all the bits are zero, Parity if the number of set bits is even. Compare will set the Equal flag if the two registers have the same bit pattern. The Greater than flag is set if S is greater than the second register. Note that when doing a compare signed/unsigned is not taken into account, the two registers are treated as if they contain two unsigned values.

NB: This might change to instead compare just the X and Y register.

Stack Manipulation

When dealing with the stack, a pair of registers will be moved to or from 'the stack' and the SP updated to reflect the changed address.

The registers A and B are paired, as are the registers C and D. Effectively, the stack works on 16 bit values, but due to the 8 bit data bus it requires two transfers, though this is handled via the hardware/microcode. Although still two distinct bytes, the B and D registers should be considered the more significant byte whilst A and C registers the lesser; the more significant byte will be stored at the lower address in the stack, the pair of registers are big-endian.

The Stack manipulation operations are of pattern 1111 10DR. The D bit indicates the direction; 0 for PUSH and 1 for POP. The R bit indicates the register pair; 0 for A & B and 1 for C & D.

When PUSHing B or D will go to the address of the SP, whilst A or C will go to address one less than the SP. After PUSHing, the SP will have been decremented by two.

When POPing, the same respective pairs of memory locations will be read from the same pair of registers, and the SP increased by two.

NB: I Think I might update this to allow pushing/popping the PC, this would make it very easy (hardware wise) to handle calling and returning functions

Variants
Reserved(u8)

Reserved instruction, contains the entire byte

Jump(InstructionJumpCondition)

Jump, see member doc

LoadImmediate

Load the the next byte into register AAA (PC will be incremented a second time)

Fields of LoadImmediate

aaa: u8
LoadIndirect

Load value in address indicated by the next two bytes into register AAA (PC will be incremented two more times)

Fields of LoadIndirect

aaa: u8
Save

Store value in register AAA in address indicated by the next two bytes (PC will be incremented two more times)

Fields of Save

aaa: u8
Alu(AluOperation)

ALU based operations, see member doc

Move

Move a value from register AAA to register BBB

Fields of Move

aaa: u8bbb: u8
Port

Perform I/O, see section above

Fields of Port

d: InstructionPortDirectionaaa: u8
Comp

Compare register S with register AAA, see section above

Fields of Comp

aaa: u8
Stck

Stack manipulation, see member doc

Fields of Stck

d: InstructionStckDirectionr: InstructionStckRegisterPair
Clrf

Clear the 'F' register, by setting it to 0000 0000

Halt

Stop the CPU from doing any more execution

Methods

impl Instruction

pub fn is_valid(&self) -> bool

Check if this instruction doesn't use reserved space

Examples

assert!(Instruction::Clrf.is_valid());
assert!(Instruction::Alu(AluOperation::Or).is_valid());

assert!(!Instruction::Reserved(0).is_valid());
assert!(!Instruction::Alu(AluOperation::Reserved(0b0011)).is_valid());

pub fn data_length(&self) -> usize

Get the amount of data bytes following this instruction

Examples

assert_eq!(Instruction::Clrf.data_length(), 0);
assert_eq!(Instruction::Alu(AluOperation::Or).data_length(), 0);

assert_eq!(Instruction::LoadImmediate{ aaa: 0 }.data_length(), 1);
assert_eq!(Instruction::Save { aaa: 0 }.data_length(), 2);

pub fn display<'r, 's: 'r>(
&'s self,
registers: &'r [GeneralPurposeRegister; 8]
) -> DisplayInstruction<'r>

Get proxy object implementing Display for printing instructions in assembly format

Examples

assert_eq!(Instruction::Clrf.display(&registers).to_string(),
           "CLRF");
assert_eq!(Instruction::Alu(AluOperation::Or).display(&registers).to_string(),
           "ALU OR");
assert_eq!(Instruction::Stck {
               d: InstructionStckDirection::Push,
               r: InstructionStckRegisterPair::Cd,
           }.display(&registers).to_string(),
           "STCK PUSH C&D");

assert_eq!(Instruction::Reserved(0b1111_0000).display(&registers).to_string(),
           "0b1111_0000");

impl Instruction

pub fn from_str(
s: &str,
registers: &[GeneralPurposeRegister; 8]
) -> Result<Instruction, ParseInstructionError>

Parse assembly instruction format

The input string must be ASCII and contain no vertical whitespace

Examples

assert_eq!(Instruction::from_str("JMPL", &registers),
           Ok(Instruction::Jump(InstructionJumpCondition::Jmpl)));

assert_eq!(Instruction::from_str("LOAD IND B", &registers),
           Ok(Instruction::LoadIndirect { aaa: 0b101 }));

assert_eq!(Instruction::from_str("ALU SOR RIGHT ASF", &registers),
           Ok(Instruction::Alu(AluOperation::ShiftOrRotate {
               d: AluOperationShiftOrRotateDirection::Right,
               tt: AluOperationShiftOrRotateType::Asf,
           })));